requirements.txt
python_dotenv == 0.19.0
spacy == 3.2.3
